# value_count Aggregation optimization