We found some problems during the test.

Data: 200Million docs， 1 shard，0 replica。


Hit | avg | sum | value_count
-- | -- | -- | --
20k | 38ms | 33ms | 63ms
200k | 127ms | 125ms | 334ms
2Million | 789ms | 729ms | 3.176s
20Million | 4.2s | 3.239s | 22.787s
200Million(100%) | 21s | 22s | 154.917s


The performance of avg, sum and other is very close when performing statistics, but the performance of value_count has always been poor, even not on an order of magnitude. Based on some common-sense knowledge, we think that value_count and sum are similar operations, and the time consumed should be the same. Therefore, we have discussed the agg of value_count.

The principle of counting in es is to traverse the field of each document. If the field is an ordinary value, the count value is increased by 1. If it is an array type, the count value is increased by n. However, the problem lies in traversing each document and taking out the field, which changes from disk to an object in the Java language. We summarize its current problems with Elasticsearch as:

- Number cast to string overhead, and GC problems caused by a large number of strings
- After the number type is converted to string, sorting and other unnecessary operations are performed


Here is the proof of type conversion overhead.

```
// Java long to string source code, getChars is very time-consuming.
public static String toString(long i) {
        int size = stringSize(i);
        if (COMPACT_STRINGS) {
            byte[] buf = new byte[size];
            getChars(i, size, buf);
            return new String(buf, LATIN1);
        } else {
            byte[] buf = new byte[size * 2];
            StringUTF16.getChars(i, size, buf);
            return new String(buf, UTF16);
        }
}   
```


test type | average | min | max | sum
-- | -- | -- | -- | --
double->long | 32.2ns | 28ns | 0.024ms | 3.22s
long->double | 31.9ns | 28ns | 0.036ms | 3.19s
long->String | 163.8ns | 93ns | 1921ms | 16.3s



https://github.com/elastic/elasticsearch/issues/36752 The program heat map shows that the toString time is particularly serious. 

## optimization

Our optimization code is actually very simple. It is to manage different types separately, instead of uniformly converting to string unified processing. We added type identification in ValueCountAggregator, and made special treatment for number and geopoint types to cancel their type conversion. Because the string type is reduced and the string constant is reduced, the improvement effect is very obvious.

## result



Hit | avg | sum |  value_countdouble before| value_countdouble after| value_countkeyword before | value_countkeyword after | value_countgeo_point before | value_countgeo_point after
-- | -- | -- | -- | -- | -- | -- | -- | --
20k | 38ms | 33ms | 63ms | 26ms | 30ms | 30ms | 38ms | 15ms
200k | 127ms | 125ms | 334ms | 78ms | 116ms | 99s | 278ms | 31ms
2Million | 789ms | 729ms | 3.176s | 439ms | 348ms | 386ms | 3.365s | 178ms
20Million | 4.2s | 3.239s | 22.787s | 2.7s | 2.5s | 2.6s | 25.192s | 1.278s
200Million(100%) | 21s | 22s | 154.917s | 18.99s | 19s | 20s | 168.971s | 9.093s

- The results are more in line with common sense. valuecount is about the same as avg, sum, etc., or even lower than these. Previously, valuecount was much larger than avg and sum, and it was not even an order of magnitude when the amount of data was large.
- When calculating numeric types such as double and long, the performance is improved by about 8 to 9 times; when calculating the geo_point type, the performance is improved by 18 to 20 times.