@xjtushilei 2020-04-08
> Nice!
> 
> @elasticmachine, ok to test.
> 
> @xjtushilei, could you sign the [CLA](https://www.elastic.co/contributor-agreement) so we can get this in and work on it?

i have signed  the CLA with Email "xunjian.sl@alibaba-inc.com"，the email is Email of GIT submitted records.
maybe it is not my github account  email?