@xjtushilei 2020-04-08
> @xjtushilei, looks like Jenkins hit a compile error. Are you ok to fix it?

I tested it in 6.x branch with no compilation errors , and the latest master branch has not been compiled.

When I tried in the master branch, ". / gradlew idea" was too slow. Could you fix it ?