@nik9000 2020-04-08
> When I tried in the master branch, ". / gradlew idea" was too slow. Could you fix it ?

I think that `./gradlew idea` has been turned off in master. I'm probably the not the right person to help because I use Eclipse, but `CONTRIBUTING.md` says to import it as a gradle project. If you don't want to go through with all that I'd be happy to finish the PR up for you, but if you want to keep contributing it is probably worth it to get gradle resolved.

> Today, I have added this email to my GitHub account email.

One of those things did the trick. I never remember all of the rules for the CLA system. Thanks for making those changes! It is all good on my side now.