@xjtushilei 2020-04-10
> @xjtushilei, I've pushed a fix for the failing test. That test actually shows that this is a breaking change but in a super esoteric way: if you run with a script the _input_ will change from a `String` to the appropriate type. In the case of the issue in comes back as some kind of number. I'll still backport it to 7.8.0 and add a breaking change note. I don't believe it is super likely to break folks so I'm ok making a breaking change in a minor for this.

That sounds cool.