@nik9000 2020-04-10
OK! Jenkins has approved this so I've merged and am backporting. I took some liberties turning your PR description into a commit [message](https://github.com/elastic/elasticsearch/commit/8e8ce967757929b1f71e6a9ce4cbfbf60407867b). You should see a backport PR sometime in the next hour or so.